change User attribute type

+++ src/main/java/com/icbt/bumble_bee/dto/UserDto.java
@@ -26,7 +26,7 @@ public class UserDto implements SuperDto {
     private String nic;
     private Date dateOfBirth;
     private Gender gender;
-    private int budget ;
+    private double budget ;
     private Date createdDate;
     private boolean isDeleted;
